backend/requirements.txt
flask
flask-cors
python-dotenv
requests
ollama
PyPDF2
python-docx
